pyproject.toml
[project]
name = "parksence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.3",
    "onnxslim>=0.1.87",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.55.0",
    "ultralytics>=8.4.21",
    "watchdog>=6.0.0",
]
